Change SpecialKey and Conceal to fgcolor=60.
SpecialKey is used for key names (like the names printed by ":map") and
unprintable characters. I use ":set list lcs=tab:\|\ " to display
indentations, and lcs chars use this highlight group. The default color
was too strong and interfered with editing. The color I wanted (236)
made key names too dim. :\ (if only lcs used a separate highlight
group). Foreground color 60 was chosen as a compromise.

Conceal is used by the indentLine plugin, so it was set to the same
value as SpecialKey (it's also used by the conceal command).